refactor: web-assests in own namespace, drop atom used
web-assets is a hash-map of useful service and content links that are shared
throughout the landing page

Separate this large hash-map into its own namespace for a nice logical
separation and to simplify the landing-page namespace

Drop the use of the atom as the web-assets data is not dynamically modified at
runtime.  It is only changed at develop time.